Configure CORS in planning server
